+++ src/backend-commons/modules/cdoc-playlist-server.module.ts
@@ -73,6 +73,67 @@ export abstract class CommonDocPlaylistServerModule<R extends CommonDocRecord, F
             });
     }
 
+    public static configureCsvPlaylistServerRoutes(app: express.Application, apiPrefix: string,
+                                                cdocPlaylistServerModule: CommonDocPlaylistServerModule<CommonDocRecord,
+                                                    CommonDocSearchForm,
+                                                    CommonDocSearchResult<CommonDocRecord, CommonDocSearchForm>,
+                                                    CommonDocDataService<CommonDocRecord, CommonDocSearchForm,
+                                                        CommonDocSearchResult<CommonDocRecord, CommonDocSearchForm>>>,
+                                                backendConfig: any | CommonBackendConfigType<CommonKeywordMapperConfigType, CacheConfig>) {
+        // use own wrapper for search
+        const exportConfig: CommonDocPlaylistExporterConfig = {
+            maxAllowed: backendConfig.playlistExportMaxM3uRecordAllowed
+        };
+        if (exportConfig.maxAllowed <= 0) {
+            console.warn('SKIP route csvplaylist NOT Enabled: playlistExportMaxM3uRecordAllowed=',
+                exportConfig.maxAllowed);
+            return;
+        }
+
+        console.log('configure route ' + cdocPlaylistServerModule.getApiId() + 'export/csvplaylist:', apiPrefix + '/:locale/'
+            + cdocPlaylistServerModule.getApiId() + 'export/csvplaylist');
+        app.route(apiPrefix + '/:locale/' + cdocPlaylistServerModule.getApiId() + 'export/csvplaylist')
+            .all(function(req, res, next) {
+                if (req.method !== 'GET') {
+                    return next('not allowed');
+                }
+                return next();
+            })
+            .get(function(req, res, next) {
+                const searchForm = this.getDataService().newSearchForm(req.query);
+                if (!cdocPlaylistServerModule.isSearchFormValid(searchForm)) {
+                    console.warn('form invalid');
+                    res.send('');
+                    return next();
+                }
+                try {
+                    cdocPlaylistServerModule.playlistExporter.exportCsvPlaylist(exportConfig, searchForm).then(playlist => {
+                        if (playlist === undefined) {
+                            console.log('csvplaylist not fullfilled');
+                            res.status(403);
+                            res.send('');
+                            return next();
+                        }
+
+                        res.set({
+                            'Content-Type': 'application/csv',
+                            'Content-Disposition': 'attachment; filename=playlist.csv'
+                        });
+                        res.status(200);
+                        res.send(playlist);
+                        return next();
+                    }).catch(reason => {
+                        console.error('csvplaylist not fullfilled:', reason);
+                        res.status(403);
+                        return next('not found');
+                    });
+                } catch (error) {
+                    console.error('error thrown: ', error);
+                    return next('not found');
+                }
+            });
+    }
+
     public constructor(protected dataService: D, protected playlistExporter: CommonDocPlaylistExporter<R, F, S, D>) {
     }